+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-23T15:59:52.215Z",
+    "2026-02-23T18:10:15.160Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {