+++ packages/orq-rc/RELEASES.md
@@ -7179,3 +7179,13 @@ Based on:
 - [typescript v4.4.0-rc.35] packages/orq-rc
 ### Releases
 - [NPM v4.4.0-rc.35] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.35 - packages/orq-rc
+
+## 2026-02-23 18:13:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.725.0 (2.836.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.4.0-rc.36] packages/orq-rc
+### Releases
+- [NPM v4.4.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.36 - packages/orq-rc